# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending_acceptance: 'pending_acceptance',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  transitioned: 'transitioned',
  expired: 'expired',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description: string;
  /** @minLength 1 */
  injectedBy: string;
  /** @nullable */
  authorityScope?: string | null;
}

export interface Task {
  id: string;
  title: string;
  description: string;
  injectedBy: string;
  /** @nullable */
  authorityScope: string | null;
  status: TaskStatus;
  createdAt: string;
}

export interface TaskAcceptanceInput {
  /** @minLength 1 */
  actorId: string;
  /** @nullable */
  contextNote?: string | null;
}

export interface TaskAcceptance {
  id: string;
  taskId: string;
  actorId: string;
  acceptedAt: string;
  /** @nullable */
  contextNote: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type Provenance = typeof Provenance[keyof typeof Provenance];


export const Provenance = {
  observed: 'observed',
  reviewed: 'reviewed',
  reported: 'reported',
} as const;

export interface TaskCompletionInput {
  /** @minLength 1 */
  actorId: string;
  provenance: Provenance;
  /** @minLength 1 */
  claimText: string;
  /** @nullable */
  sourceReference?: string | null;
}

export interface TaskCompletion {
  id: string;
  taskId: string;
  actorId: string;
  provenance: Provenance;
  claimText: string;
  /** @nullable */
  sourceReference: string | null;
  reportedAt: string;
}