# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tasks:
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      description: Creates a task. Status always starts as pending_acceptance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}/accept:
    post:
      operationId: acceptTask
      tags: [tasks]
      summary: Accept a task
      description: >-
        Creates a task_acceptances row and moves the task's status to
        accepted. Fails if the task already has an acceptance record,
        regardless of which actor is attempting to accept it — ownership is
        first-come, first-served and can never be silently overwritten.
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskAcceptanceInput"
      responses:
        "201":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAcceptance"
        "404":
          description: Task or actor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Task already has an acceptance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{taskId}/complete:
    post:
      operationId: reportCompletion
      tags: [tasks]
      summary: Report a task completion
      description: >-
        Creates a task_completions row. provenance must be exactly one of
        observed, reviewed, or reported — no other value, casing, or missing
        value is accepted. The call is rejected outright (no row created) if
        provenance is missing or invalid. On success, the task's status
        moves to completed.
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompletionInput"
      responses:
        "201":
          description: Completion recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompletion"
        "400":
          description: Missing or invalid provenance, or other invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task or actor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    TaskStatus:
      type: string
      enum:
        - pending_acceptance
        - accepted
        - in_progress
        - completed
        - transitioned
        - expired
    TaskInput:
      type: object
      required:
        - title
        - description
        - injectedBy
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        injectedBy:
          type: string
          minLength: 1
        authorityScope:
          type: ["string", "null"]
    Task:
      type: object
      required:
        - id
        - title
        - description
        - injectedBy
        - authorityScope
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        injectedBy:
          type: string
        authorityScope:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/TaskStatus"
        createdAt:
          type: string
          format: date-time
    TaskAcceptanceInput:
      type: object
      required:
        - actorId
      properties:
        actorId:
          type: string
          minLength: 1
        contextNote:
          type: ["string", "null"]
    TaskAcceptance:
      type: object
      required:
        - id
        - taskId
        - actorId
        - acceptedAt
        - contextNote
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        actorId:
          type: string
        acceptedAt:
          type: string
          format: date-time
        contextNote:
          type: ["string", "null"]
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    Provenance:
      type: string
      enum:
        - observed
        - reviewed
        - reported
    TaskCompletionInput:
      type: object
      required:
        - actorId
        - provenance
        - claimText
      properties:
        actorId:
          type: string
          minLength: 1
        provenance:
          $ref: "#/components/schemas/Provenance"
        claimText:
          type: string
          minLength: 1
        sourceReference:
          type: ["string", "null"]
    TaskCompletion:
      type: object
      required:
        - id
        - taskId
        - actorId
        - provenance
        - claimText
        - sourceReference
        - reportedAt
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        actorId:
          type: string
        provenance:
          $ref: "#/components/schemas/Provenance"
        claimText:
          type: string
        sourceReference:
          type: ["string", "null"]
        reportedAt:
          type: string
          format: date-time