# lib/api-zod/src/generated/types/taskCompletion.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Provenance } from './provenance';

export interface TaskCompletion {
  id: string;
  taskId: string;
  actorId: string;
  provenance: Provenance;
  claimText: string;
  /** @nullable */
  sourceReference: string | null;
  reportedAt: Date;
}