# lib/api-zod/src/generated/types/taskCompletionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Provenance } from './provenance';

export interface TaskCompletionInput {
  /** @minLength 1 */
  actorId: string;
  provenance: Provenance;
  /** @minLength 1 */
  claimText: string;
  /** @nullable */
  sourceReference?: string | null;
}